Fix Windows-8.1 e2e test failures.
We install google-compute-engine-metadata-scripts to simplify scheduling
startup tasks on Windows VMs.

This package creates and schedules a task `GCEStartup` that later reads
the value of `windows-startup-script-ps1` stored in metadata.

For some reason, on Windows-10 VMs this task assigned a wrong
activation value and therefore is not invoked on startup.

After reinstalling the package this issue is solved. See this [PR](
https://github.com/GoogleCloudPlatform/compute-image-tools/pull/1165
) for details.

However, it corrupts the activation time of `GCEStartup` on Windows-8.1
VMs. Therefore we need to check the Windows version before reinstalling
this package.

One small note, we need to compare the version with a string value
instead of a number, otherwise this check will fail for numbers that can
be considered as integers, for example, 10.0.