Start using N2 instances for SLES import.
Some SLES tests has started failing since the end of July because of
timeouts. SLES import uses E2 instances that can be not optimal for the
import processes. The idea is to stay consistent with other Linux
distributions and to use N1 instances.

On the other side, as N2 instances provide better performance than N1 we
can update the import for SLES to start using N2 and check how it goes.
It will be a good starting point before updating machine types for other
distributions and tools.