Attach disks to the worker after it is created.
Import / Export tools are built on the assumption that worker disks of
Linux workers will always receive the name `/dev/sda`. It worked well up
to Debian-10 workers. Since Debian-11 this assumption has stopped
working and the names are assigned quite randomly. It leads to the cases
when the worker disk receives the name `/dev/sdb` that breaks the
behaviour of our tools.

This commit updates RHEL workflows (where we already use Debian-11
workers) in the way that input disks are attached after a worker
instance is created, so the worker disk of this instance will always
receive the name `/dev/sda`.

The solution is temporary, in the future we will implement
auto-detection logic for worker disks.