Image Import: Fix inspection module tests. (#58)
The inspection module tests depend on Google Cloud official images, when
some images are deprecated and removed the tests start failing.

In the context of this PR we are updating the tests to start using
families instead of direct links to images for SLES.

In the future we should stop using official images for testing as they
make our tests fragile.