Fix SLES PAYG E2E tests.
This is an update to the previous PR
https://github.com/GoogleCloudPlatform/compute-image-import/pull/60.
Zypper requires an active subscription to install updates. However, if
the image is intended to be used on GCE with a PAYG license, then it
doesn't need to have a subscritpion. And therefore we can request Zypper
updates only for BYOL images.