Nested virtualization. (#75)
The libguestfs library that is used internally in the tools runs a VM
during performing its operations.

Therefore when a worker VM runs this libguestfs VM, the libguestfs VM is
created as a **nested** VM.

By default, GCP VMs don't enable hardware virtualization for nested VMs
until a special flag `--enable-nested-virtualization` is provided.

See
https://cloud.google.com/compute/docs/instances/nested-virtualization/enabling
for details.