Revert gcs bucket update (#168)
* Revert "Remove soft deletion when GCS bucket is provided as an arg (#159)"

This might be risky.
In case customers are using the scratch bucket to store their own data and rely on the soft deletion policy.
New buckets are created by gcloud without soft-deletion.

This reverts commit faeea21ed3e230f2d4fe3effb6085ab94e45abc9.

* Revert "Reapply "Update bucket attributes in order to disable soft delete (#155)" (#156) (#157)"

This might be risky.
In case customers are using the scratch bucket to store their own data and rely on the soft deletion policy.

This reverts commit 823d6d2af4c6ac4a46df0a32cfdc45a4f089bc4e.